Moved consumer group controller
